cleaned up emu cycles triggers to be done on a per instruction basis rather than by dispirate calls

+++ internal/emu/context.go
@@ -26,7 +26,7 @@ func NewContext() *Context {
 	}
 }
 
-func (c *Context) EmuCycle(i int) {
+func (c *Context) EmuCycle(i uint8) {
 	for range i {
 		for range 4 {
 			c.ticks++
